Actualizo lógica de autenticación

+++ src/main/java/pe/edu/upc/trabajoparcial/repositories/ProductoRepository.java
@@ -4,20 +4,20 @@ import org.springframework.stereotype.Repository;
 import pe.edu.upc.trabajoparcial.entities.Producto;
 
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Repository
 public interface ProductoRepository extends JpaRepository<Producto, Integer> {
 
-    // Buscar productos por nombre
     List<Producto> findByNombreContainingIgnoreCase(String nombre);
 
-    // Buscar productos por categoría
     List<Producto> findByCategoriaNombreContainingIgnoreCase(String categoria);
 
-    // Buscar productos por rango de precio
-    List<Producto> findByPrecioBetween(Float minPrecio, Float maxPrecio);
+    // **CORRECCIÓN**: usar "Between"
+    List<Producto> findByPrecioBetween(BigDecimal minPrecio, BigDecimal maxPrecio);
 
-    // Buscar productos de un vendedor
     List<Producto> findByClienteIdCliente(Integer idCliente);
+
+
 }
